core: create fs module and move fs commands there
